% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmq.R
\name{r2r.call}
\alias{r2r.call}
\title{Send an R object to a listening RMQ server.}
\usage{
r2r.call(msg, addr = rmq.default.addr)
}
\description{
\code{r2r.call()} is the client counter-part to \code{r2r.server()}
}
\examples{
\dontrun{
  x=list()
  x$arg=c(1,2,3)
  x$f = function(y) { sum(y) }
  r2r.call(x)
}
}
\seealso{
\code{r2r.server},
  \code{rmq.server},
  \code{rmq.call}
}

