% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmq.R
\name{read.msgpack.frame}
\alias{read.msgpack.frame}
\title{create an R object from the head of a stream of framed binarray msgpack bytes}
\usage{
read.msgpack.frame(x, byteOffset)
}
\arguments{
\item{x}{A raw byte vector of one or more bin-array-framed-msgpack objects.}

\item{byteOffset}{specifies which object to return, namely the one starting at x[byteOffset]}
}
\value{
A list with two elements. The first is the next byteOffset that should be used to read the next object (the one after the one returned from the current call); the second is the R object represented by the frame starting at the supplied byteOffset.
}
\description{
Given a vector of raw bytes written in framed-msgpack format, \code{read.msgpack.frame} converts these into an list of 1) the next byteOffset to be used for reading the next frame; and 2) the correspending reified R object obtained from converting the frame at byteOffset from msgpack into R.
}
\seealso{
\url{http://msgpack.org}

Other rmq.functions: \code{\link{from.msgpack}},
  \code{\link{r2r.call}}, \code{\link{r2r.server}},
  \code{\link{rmq.call}}, \code{\link{rmq.default.addr}},
  \code{\link{rmq.server}}, \code{\link{to.msgpack}}
}

