% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmq.R
\name{rmq.call}
\alias{rmq.call}
\title{Send a message to a listening RMQ server.}
\usage{
rmq.call(msg, addr = rmq.default.addr, timeout.msec = 5000)
}
\arguments{
\item{msg}{An R object. Can be a list. Internally this will be converted into msgpack and sent to the server.}

\item{addr}{A string of "IP:port" format. The server will bind \code{addr}, and it must be available. Defaults to \code{rmq.default.addr}, which is "127.0.0.1:9090".}

\item{timeout.msec}{A timeout value in milliseconds. A value of 0 means wait forever for a reply. It is recommended to use a small finite timeout such as the 5 second default, because there is no other way to interrupt the \code{rmq.call()} while it is waiting on the network. Issuing ctrl-c (SIGINT) in particular will not interrupt the \code{rmq.call()} in progress.}
}
\value{
The return value is the response from the rmq server to the given msg.
}
\description{
Send a message to a listening RMQ server.
}
\examples{
\dontrun{
  rmq.call(msg, addr="10.0.0.1:7777", timeout.msec = 1000)
}
}
\seealso{
Other rmq.functions: \code{\link{from.msgpack}};
  \code{\link{r2r.call}}; \code{\link{r2r.server}};
  \code{\link{rmq.default.addr}}; \code{\link{rmq.server}};
  \code{\link{to.msgpack}}
}

