% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmq.R
\docType{data}
\name{rmq.default.addr}
\alias{rmq.default.addr}
\title{The default address bound by \code{rmq.server}.}
\format{\preformatted{ chr "127.0.0.1:9090"
}}
\usage{
rmq.default.addr
}
\description{
The default address bound by \code{rmq.server}.
}
\seealso{
Other rmq.functions: \code{\link{from.msgpack}};
  \code{\link{r2r.server}}; \code{\link{rmq.call}};
  \code{\link{rmq.server}}; \code{\link{to.msgpack}}
}
\keyword{datasets}

