% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmq.R
\name{rmq.server}
\alias{rmq.server}
\title{Start an RMQ server, listening on specified IP and port.}
\usage{
rmq.server(handler, addr = rmq.default.addr)
}
\arguments{
\item{handler}{A handler R function taking a single argument}

\item{addr}{A string of "IP:port" format. The server will bind \code{addr}, and it must be available. Defaults to \code{rmq.default.addr}, which is "127.0.0.1:9090".}
}
\value{
No return value. Blocks forever listening and calling the handler function when a request arrives. Ctrl-c will interrupt the server and shut it down. Call \code{rms.server()} again to re-start the server.
}
\description{
Start an RMQ server, listening on specified IP and port.
}
\examples{
\dontrun{
   ## the simplest echo server
   rmq.server(handler=function(msg) {msg}, addr="10.0.0.1:7777")

  ## This second example is a simple handler
  ## that echos the input it receives, and adds a few other things.
 handler = function(x) {
   print("handler called back with argument x = ")
   print(x)
   reply=list()
   reply$hi = "there!"
   reply$yum = c(1.1, 2.3)
   reply$input = x
   reply
 }
 r = rmq.server(handler, addr=rmq.default.addr)
}
}
\seealso{
Other rmq.functions: \code{\link{from.msgpack}};
  \code{\link{r2r.server}}; \code{\link{rmq.call}};
  \code{\link{rmq.default.addr}}; \code{\link{to.msgpack}}
}

