% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmq.R
\name{to.msgpack}
\alias{to.msgpack}
\title{serialize an R object to raw msgpack bytes}
\usage{
to.msgpack(x)
}
\arguments{
\item{x}{An R object to be serialized. Lists, numeric vectors, raw vectors, and string vectors are supported.}
}
\value{
A raw byte vector containing the msgpack serialized object.
}
\description{
Given an R object, \code{to.msgpack} will convert that object to a vector of raw bytes written in msgpack format.
}
\details{
Lists, numeric vectors, integer vectors, string vectors, and raw byte vectors are supported.
}
\examples{
x=list()
   x$hello = "rmq"
   raw=to.msgpack(x)
   y=from.msgpack(raw)
   ## y and x should be equal
}
\seealso{
\url{http://msgpack.org}

Other rmq.functions: \code{\link{from.msgpack}};
  \code{\link{r2r.call}}; \code{\link{r2r.server}};
  \code{\link{rmq.call}}; \code{\link{rmq.default.addr}};
  \code{\link{rmq.server}}
}

