% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmq.R
\name{r2r.call}
\alias{r2r.call}
\title{Send an R object to a listening RMQ server.}
\usage{
r2r.call(msg, addr = rmq.default.addr)
}
\arguments{
\item{msg}{An R object. Can be a list. Internally this will be serialized using \code{serialize}, then converted into a msgpack binary array and sent to the server.}

\item{addr}{A string of "IP:port" format. The server will bind \code{addr}, and it must be available. Defaults to \code{rmq.default.addr}, which is "127.0.0.1:9090".}
}
\description{
\code{r2r.call()} is the client counter-part to \code{r2r.server()}
}
\examples{

\dontrun{ 
  x=list()
  x$arg=c(1,2,3)
  x$f = function(y) { sum(y) }
  r2r.call(x)
}
}
\seealso{
Other rmq.functions: \code{\link{from.msgpack}},
  \code{\link{r2r.server}},
  \code{\link{read.msgpack.frame}},
  \code{\link{read.ndjson}}, \code{\link{rmq.call}},
  \code{\link{rmq.default.addr}}, \code{\link{rmq.server}},
  \code{\link{to.msgpack}}
}

