% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmq.R
\docType{package}
\name{rmq}
\alias{rmq}
\alias{rmq-package}
\title{R Messaging and Queuing: msgpack2 serialization and RPC over websockets}
\description{
RMQ lets you do msgpack2 encoding and decoding, and provides a
websocket based remote procedure call (RPC) mechanism.
}
\details{
The basic server and client functions are \code{\link{rmq.server}}
and \code{\link{rmq.call}}. The client and server communicate
internally by encoding and decoding to msgpack2 bytes on the wire.
Msgpack2 is the upgraded msgpack spec that distinguishes between
blobs and utf8 strings.

Client and server use the websocket protocol which means the
server can be accessed from the broswer-based javascript, and
the calls will go through firewalls without issue. The gorilla websocket
implementation supports TLS certificates for security. A user
supplied R function is invoked by the server to handle each
incoming client connection.

You can also make use of \code{\link{to.msgpack}} and
 \code{\link{from.msgpack}} diretly for situations that do not
require remote procedure call or websockets.
}
\references{
\url{https://github.com/glycerine/rmq}, \url{http://msgpack.org}
}

