% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmq.R
\docType{package}
\name{rmq}
\alias{rmq}
\alias{rmq-package}
\title{R Messaging and Queuing: msgpack2 serialization and RPC over websockets}
\description{
RMQ lets you do msgpack2 encoding and decoding, and provides a
websocket based remote procedure call (RPC) mechanism.
}
\details{
The basic server and client functions are \code{\link{rmq.server}}
and \code{\link{rmq.call}}. The client and server communicate
internally by encoding and decoding to msgpack2 bytes on the wire.
Msgpack2 is the upgraded msgpack spec that distinguishes between
blobs and utf8 strings.

Client and server use the websocket protocol which means the 
server can be accessed from the broswer-based javascript, and
the calls will go through firewalls without issue. The gorilla websocket 
implementation supports TLS certificates for security. A user
supplied R function is invoked by the server to handle each
incoming client connection.

You can also make use of \code{\link{to.msgpack}} and
 \code{\link{from.msgpack}} diretly for situations that do not
require remote procedure call or websockets.

In summary, see \code{\link{to.msgpack}} and \code{\link{from.msgpack}}
for stand-alone fast msgpack2 serialization within an R session. See 
\code{\link{rmq.server}} and \code{\link{rmq.call}} for
the basic RPC over websocket functionality that allows inter-operations
with other languages using msgpack2. The default listening address 
is defined in \code{\link{rmq.default.addr}}.
This overview of RMQ is found under the title \code{\link{rmq}}.

With a little elaboration on these
basics, we can exchange arbitrary R data and functions,
\code{\link{r2r.server}} and \code{\link{r2r.call}} for R-to-R
object transfer. 
The rmq package comes with two simple example scripts that illustrate
using these r2r functions. See \url{https://github.com/glycerine/rmq/blob/master/example-client.R}
and \url{https://github.com/glycerine/rmq/blob/master/example-server.R}
in the source package for a simple working system to get started from.
}
\references{
\url{https://github.com/glycerine/rmq}, \url{http://msgpack.org}
}

