% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmq.R
\name{rmq.server}
\alias{rmq.server}
\title{Start an RMQ server, listening on specified IP and port.}
\usage{
rmq.server(handler, addr = rmq.default.addr)
}
\arguments{
\item{handler}{A handler R function taking a single argument}

\item{addr}{A string of "IP:port" format. The server will bind \code{addr}, and it must be available. Defaults to \code{rmq.default.addr}, which is "127.0.0.1:9090".}
}
\value{
No return value. Blocks forever listening and calling the handler function when a request arrives. Ctrl-c will interrupt the server and shut it down. Call \code{rms.server()} again to re-start the server.
}
\description{
Start an RMQ server, listening on specified IP and port.
}
\examples{
\dontrun{ 
   ## a) the simplest echo server - run this in R session #1. Assumes port
   ##    9090 (defined in package variable rmq.default.addr) is available locally.
   rmq.server(handler=function(msg) {msg})

  ## b) This second example is a simple handler
  ## that echos the input it receives, and adds a few other things.
  ## This would also be in R session #1, as an alternative
  ## to a) above.
 handler = function(x) {
   print("handler called back with argument x = ")
   print(x)
   reply=list()
   reply$hi = "there!"
   reply$yum = c(1.1, 2.3)
   reply$input = x
   reply
 }
 r = rmq.server(handler)

 ## c) lastly the client call. In R session #2. You'll
 ## always need to run c) after first starting the 
 ## the server using a) or b) above in a separate
 ## R session.
 rmq.call("hello rmq!")

 ## d) illustrate how the client call can pass complex
 ## nested list structured data.
 monster=list()
 eyes=list()
 eyes$description = c("red","glowing")
 monster$eyes = eyes
 monster$measurements = c(34, 22, 33)

 ## finally, send the monster to the server.
 rmq.call(monster)

}


}
\seealso{
Other rmq.functions: \code{\link{from.msgpack}},
  \code{\link{r2r.call}}, \code{\link{r2r.server}},
  \code{\link{read.msgpack.frame}},
  \code{\link{read.ndjson}}, \code{\link{rmq.call}},
  \code{\link{rmq.default.addr}}, \code{\link{to.msgpack}}
}

